// src/app/types/ResTypes.ts
export interface Pagination {
  total_pages: number;
  total_records: number;
  current_page: number;
  records_per_page: number;
}
// ...
export interface LikeListRes {
    status:  boolean;
    data:    LikeListData;
    message: string;
    toast:   boolean;
}

export interface LikeListData {
    Records:    LikeListRecord[];
    Pagination: Pagination;
}

export interface LikeListPagination {
    total_pages:      number;
    total_records:    number;
    current_page:     number;
    records_per_page: number;
}

export interface LikeListRecord {
    like_by:        number;
    social_id:      number;
    comment_id:     null;
    Social:         LikeListSocial;
    isLiked:        boolean;
    isSaved:        boolean;
    isFollowing:    boolean;
    total_comments: number;
    total_likes:    number;
    total_saves:    number;
}